backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
SQLAlchemy==2.0.40
pymysql==1.1.1
python-dotenv==1.1.0
loguru==0.7.3
openai==1.75.0
anthropic==0.49.0
httpx==0.28.1
httpx-sse==0.4.0
aiohttp==3.11.18
passlib==1.7.4
python-jose==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20
pytest==8.3.5
pytest-asyncio==0.26.0
alembic==1.15.2
aiomysql==0.2.0
tqdm==4.67.1
mcp==1.0.0
psutil==7.0.0
aiosqlite==0.21.0